Move layout to layouts
