+++ src/pages/Home/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import { HeaderContent } from '../../layout'
+import { HeaderContent } from '../../layouts'
 
 export const Home: React.FC = () => (
   <HeaderContent>Welcome to We Do Takeaway</HeaderContent>
